support flat messages
Since July 2018, the devtools protocol has supported attaching to
targets with flatten=true. Then, messages are sent and received from the
target via a sessionId field on the Message type itself instead of using
sendMessageToTarget and eventReceivedFromTarget.

Add the optional field to Message, which is necessary to support the
mechanism. It's optional since not specifying a SessionID simply sends a
command to the browser itself.

While at it, make a couple of minor changes to better support generating
the code inside a module. First, follow the GO111MODULES environment
variable set by the user; it's always possible to generate to GOPATH via
GO111MODULES=on. Second, always ensure that the "out" path is absolute,
since the code relies on it. Otherwise -out=. would fail in weird ways.